added in logo added in error box before layout changers
