+++ dash_Graphs.py
@@ -8,9 +8,10 @@ import matplotlib.pyplot as plt
 import datetime as dt
 import webbrowser
 from calendar import day_name
+import ctypes  # An included library with Python install.
 
 
-
+from fcn_UTILS import character_removal, dataframe_chooser, Mbox
 
 ## DASH ##
 
@@ -20,33 +21,11 @@ import dash_core_components as dcc
 import dash_html_components as html
 from dash.dependencies import Input, Output
 
-
-def character_removal(string_to_filter): 
-    chars_to_remove = ['[', ']', '\'']
-    filtered_list = []
-    #determine if it is a list 
-    if isinstance(string_to_filter, list): #if it is a list 
-        for x in range(0, len(string_to_filter)): #iterate through lenghth of list
-            for char in chars_to_remove: #filter through characters to remove
-                string_to_filter[x] = string_to_filter[x].replace(char, '') #remove each character
-            filtered_list.append(string_to_filter[x]) #append each filtered word to a new
-        return filtered_list #return it 
-        
-    else: #not a list
-        for char in chars_to_remove: #filter through chars to remove
-            string_to_filter = string_to_filter.replace(char, '') #remove the char by replacing it with nothing
-    return string_to_filter #return it 
-
-def dataframe_chooser(Daily_Interval_Data, chosen_site): 
-    """function to dynamically slice and create a new dataframe from given dataframes"""
-    Months = ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December'] #list of the months to access things later
-    dynamically_created_dataframe = pd.concat([Daily_Interval_Data[0].loc[:, chosen_site], Daily_Interval_Data[1].loc[:, chosen_site],Daily_Interval_Data[2].loc[:, chosen_site],
-                            Daily_Interval_Data[3].loc[:, chosen_site],Daily_Interval_Data[4].loc[:, chosen_site],Daily_Interval_Data[5].loc[:, chosen_site],
-                            Daily_Interval_Data[6].loc[:, chosen_site],Daily_Interval_Data[7].loc[:, chosen_site],Daily_Interval_Data[8].loc[:, chosen_site],
-                            Daily_Interval_Data[9].loc[:, chosen_site],Daily_Interval_Data[10].loc[:, chosen_site],Daily_Interval_Data[11].loc[:, chosen_site]], axis = 1) # append all DF's into a single dataframe YES THIS IS JANNKY I WILL FIX IT LATER 
-    dynamically_created_dataframe.columns = Months #make the column names line up 
-
-    return dynamically_created_dataframe
+from functools import lru_cache
+...
+@lru_cache(maxsize=32)
+def get_config_file(*args):
+    ...
 
 
 
@@ -88,18 +67,20 @@ def Dash_App(Daily_Interval_Data, Weekly_Interval_Data, Monthly_Sum):
     app = dash.Dash(__name__, external_stylesheets=external_stylesheets)
 
     app.layout = html.Div([
-        html.P('Some words go here'),
+        html.Img(src=app.get_asset_url('NEW_LOGO.jpg')), #display NEW LOGO
+        html.P('Historical Energy Load Anaylsis Tool'), #quick graph
         dcc.Dropdown(id = 'Drop_Down_menu',
             options=[{'label':name, 'value':name} for name in names],
             value = names[0],
             multi=False #do not allow multiple 
             ), 
-        dcc.Graph(id='graph-with-slider')
+        dcc.Graph(id='daily_graph'), 
+        dcc.Graph(id='weekly_graph'), 
     ])
 
 
     @app.callback(
-        Output('graph-with-slider', 'figure'),
+        Output('daily_graph', 'figure'),
         [Input('Drop_Down_menu', 'value')])
     def update_figure(selected_name):
         #filter the names 
@@ -107,8 +88,25 @@ def Dash_App(Daily_Interval_Data, Weekly_Interval_Data, Monthly_Sum):
 
         ### DYNAMICALLY CREATE DATAFRAME TO SHOW ALL MONTHS ###
         dataframe_to_plot = dataframe_chooser(Daily_Interval_Data, chosen_site) #dynamically create dataframes to plot entire year of chosen 
-        
-        fig = dataframe_to_plot.iplot(kind = 'line', xTitle='Site', yTitle='Consumption (kWh)', title = 'Consumption', asFigure = True) #create a list of figures 
+        try: 
+            fig = dataframe_to_plot.iplot(kind = 'line', xTitle='Site', yTitle='Consumption (kWh)', title = 'Consumption', asFigure = True) #create a list of figures
+        except KeyError: #https://github.com/santosjorge/cufflinks/issues/180
+            Mbox('PLOT ERROR', 'Dash has encountered an error. Please select another site, and try again', 1)
+
+        return fig
+    @app.callback(
+        Output('weekly_graph', 'figure'),
+        [Input('Drop_Down_menu', 'value')])
+    def update_figure1(selected_name):
+        #filter the names 
+        chosen_site = character_removal(selected_name) #this sanitises the chosen input, which is gleaned from Dash Selection
+
+        ### DYNAMICALLY CREATE DATAFRAME TO SHOW ALL MONTHS ###
+        dataframe_to_plot = dataframe_chooser(Weekly_Interval_Data, chosen_site) #dynamically create dataframes to plot entire year of chosen 
+        try: 
+            fig = dataframe_to_plot.iplot(kind = 'line', xTitle='Site', yTitle='Consumption (kWh)', title = 'Consumption', asFigure = True) #create a list of figures
+        except KeyError: #https://github.com/santosjorge/cufflinks/issues/180
+            Mbox('PLOT ERROR', 'Dash has encountered an error. Please select another site, and try again', 1)
 
         return fig
